% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{association_mutations}
\alias{association_mutations}
\title{List providing mutations as binary variables and signatures assignments for 656 breast cancer patients}
\format{
List providing mutations as binary variables and signatures assignments for 656 breast cancer patients from ICGC (https://dcc.icgc.org/)
}
\source{
ICGC data portal (https://dcc.icgc.org/).
}
\usage{
data(association_mutations)
}
\value{
Mutations and SBS assignments
}
\description{
Mutations and SBS assignments for 656 breast tumors from ICGC.
}
