% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{association_survival}
\alias{association_survival}
\title{List providing clinical data and signatures assignments for 359 pancreatic cancer patients}
\format{
List providing clinical data and signatures assignments for 359 pancreatic cancer patients from ICGC (https://dcc.icgc.org/)
}
\source{
ICGC data portal (https://dcc.icgc.org/).
}
\usage{
data(association_survival)
}
\value{
Clinical data and SBS assignments
}
\description{
Clinical data and SBS assignments for 359 pancreatic tumors from ICGC.
}
