% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/associations.prognosis.R
\name{associationPrognosis}
\alias{associationPrognosis}
\title{associationPrognosis}
\usage{
associationPrognosis(clinical_data, signatures)
}
\arguments{
\item{clinical_data}{Matrix with two columns proving survival times (first column) and status (second column).}

\item{signatures}{Matrix with the estimated exposures to signatures (columns) for each patient (rows).}
}
\value{
A vector of coefficients resulting from regularized Cox regression.
}
\description{
Perform the estimation of mutational signatures associated to prognosis.
}
\examples{
data(association_survival)
set.seed(12345)
clinical_data = association_survival$clinical_data
normalized_alpha = association_survival$normalized_alpha
prognosis_associations = associationPrognosis(clinical_data = clinical_data, signatures = normalized_alpha)

}
