% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{background}
\alias{background}
\title{Germline replication error}
\format{
Vector of rates
}
\source{
Nat Genet. 2016 Feb;48(2):126-133 (https://www.nature.com/articles/ng.3469).
}
\usage{
data(background)
}
\value{
Vector of rates for the 96 trinucleotides
}
\description{
Germline replication error estimated in Rahbari, Raheleh, et al. (2016).
}
