% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{background2}
\alias{background2}
\title{COSMIC replication error}
\format{
Vector of rates
}
\source{
COSMIC database (https://cancer.sanger.ac.uk/cosmic/signatures) v3.
}
\usage{
data(background2)
}
\value{
Vector of rates for the 96 trinucleotides
}
\description{
Background replication error signature derived from COSMIC SBS5.
}
