% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bsg}
\alias{bsg}
\title{BSgenome from the Bioconductor package BSgenome.Hsapiens.1000genomes.hs37d5}
\format{
reference BSgenome
}
\source{
Bioconductor package BSgenome.Hsapiens.1000genomes.hs37d5
}
\usage{
data(bsg)
}
\value{
reference BSgenome from the Bioconductor package BSgenome.Hsapiens.1000genomes.hs37d5
}
\description{
reference BSgenome from the Bioconductor package BSgenome.Hsapiens.1000genomes.hs37d5
}
