% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cn_example_reduced}
\alias{cn_example_reduced}
\title{A reduced version of the copy number data for 5 TCGA samples in the format compatible with the import function}
\format{
Reduced versione of the counts of copy numbers in the format compatible with the import function
}
\source{
Nature. 2022 Jun;606(7916):984-991 (https://www.nature.com/articles/s41586-022-04738-6).
}
\usage{
data(cn_example_reduced)
}
\value{
Reduced versione of the counts of copy numbers for 5 tumors and 48 copy number classes in the format compatible with the import function
}
\description{
Reduced versione of the dataset of counts of copy numbers detected in TCGA tumors published in Steele, Christopher D., et al. (2022).
}
