% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.import.R
\name{get.MNV.counts}
\alias{get.MNV.counts}
\title{get.MNV.counts}
\usage{
get.MNV.counts(data)
}
\arguments{
\item{data}{a data.frame with variants having 6 columns: sample name, chromosome, start position, end position, ref, alt.}
}
\value{
A matrix with Multi-Nucleotide Variants (MNVs) counts per patient.
}
\description{
Create Multi-Nucleotide Variants (MNVs) counts matrix from input data.
}
\examples{
data(ssm560_reduced)
res = get.MNV.counts(data = ssm560_reduced)

}
