% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.import.R
\name{getCNCounts}
\alias{getCNCounts}
\title{getCNCounts}
\usage{
getCNCounts(data)
}
\arguments{
\item{data}{A data.frame with copy number data having 6 columns: sample name, chromosome, start position, end position, major CN, minor CN.}
}
\value{
A matrix with Copy Numbers (CNs) counts per patient.
}
\description{
Create Copy Numbers (CNs) counts matrix from input data.
This function has been derived from: https://github.com/UCL-Research-Department-of-Pathology/panConusig/blob/main/R/setup_CNsigs.R
}
\examples{
data(cn_example_reduced)
res <- getCNCounts(data = cn_example_reduced)

}
