% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{groups.SBS.plot}
\alias{groups.SBS.plot}
\title{groups.SBS.plot}
\usage{
groups.SBS.plot(counts, groups, normalize = TRUE, xlabels = FALSE)
}
\arguments{
\item{counts}{matrix with Single Base Substitutions (SBS) counts data.}

\item{groups}{list where names are groups labels and elements are patients labels corresponding to rownames in counts.}

\item{normalize}{boolean value; shall I normalize observed counts?}

\item{xlabels}{boolean value; shall I display x labels?}
}
\value{
A ggplot2 object.
}
\description{
Plot observed Single Base Substitutions (SBS) counts for different groups of patients.
}
\examples{
data(plot_data_examples)
counts = plot_data_examples[["groups.SBS.plot"]][["counts"]]
groups = plot_data_examples[["groups.SBS.plot"]][["groups"]]
groups.SBS.plot(counts=counts,groups=groups)

}
