% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{groupsIDPlot}
\alias{groupsIDPlot}
\title{groupsIDPlot}
\usage{
groupsIDPlot(counts, groups, normalize = TRUE, xlabels = FALSE)
}
\arguments{
\item{counts}{Matrix with Small Insertions and Deletions (ID) counts data.}

\item{groups}{List where names are groups labels and elements are patients labels corresponding to rownames in counts.}

\item{normalize}{Boolean value; shall I normalize observed counts?}

\item{xlabels}{Boolean value; shall I display x labels?}
}
\value{
A ggplot2 object.
}
\description{
Plot observed Small Insertions and Deletions (ID) counts for different groups of patients.
}
\examples{
data(plot_data_examples)
counts <- plot_data_examples[['groups.ID.plot']][['counts']]
groups <- plot_data_examples[['groups.ID.plot']][['groups']]
groupsIDPlot(counts=counts,groups=groups)

}
