% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{id_example_reduced}
\alias{id_example_reduced}
\title{A reduced version of the indel data for 3 samples in the format compatible with the import function}
\format{
Reduced version of the counts of indels in the format compatible with the import function
}
\source{
Cell Rep. 2018 Nov 27;25(9):2308-2316.e4 (10.1016/j.celrep.2018.11.014).
}
\usage{
data(id_example_reduced)
}
\value{
Reduced version of the counts of indels for 3 samples and 83 indel classes in the format compatible with the import function
}
\description{
Reduced version of the dataset of counts of indels detected in 3 samples from Osorio, Fernando G., et al. (2018).
}
