% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutational.signatures.discovery.R
\name{nmf.fit}
\alias{nmf.fit}
\title{nmf.fit}
\usage{
nmf.fit(x, beta)
}
\arguments{
\item{x}{data matrix.}

\item{beta}{signature matrix.}
}
\description{
Perform fit of NMF solution by Non-negative least squares.
}
