% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutational.signatures.discovery.R
\name{nmf.nnls}
\alias{nmf.nnls}
\title{nmf.nnls}
\usage{
nmf.nnls(x, seed, background = NULL)
}
\arguments{
\item{x}{data matrix.}

\item{seed}{initial seed for the inference.}

\item{background}{background signature.}
}
\description{
Perform NMF by Non-negative least squares.
}
