% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutational.signatures.discovery.R
\name{nmf.seed}
\alias{nmf.seed}
\title{nmf.seed}
\usage{
nmf.seed(model, target)
}
\arguments{
\item{model}{NMF model.}

\item{target}{data matrix.}
}
\description{
Initialize alpha and beta for nmf.nnls or nmf.lasso functions.
}
