% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{patients.CN.plot}
\alias{patients.CN.plot}
\title{patients.CN.plot}
\usage{
patients.CN.plot(
  cn_data_counts,
  samples = rownames(cn_data_counts),
  freq = FALSE,
  xlabels = FALSE
)
}
\arguments{
\item{cn_data_counts}{Copy Number counts matrix.}

\item{samples}{name of the samples. This should match a rownames in cn_data_counts}

\item{freq}{boolean value; shall I display rates instead of counts?}

\item{xlabels}{boolean value; shall I display x labels?}
}
\value{
A ggplot2 object.
}
\description{
Plot Copy Number (CN) counts for a set of given patients.
}
