% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{patients.MNV.plot}
\alias{patients.MNV.plot}
\title{patients.MNV.plot}
\usage{
patients.MNV.plot(
  multi_nucleotides_counts,
  samples = rownames(multi_nucleotides_counts),
  freq = FALSE,
  xlabels = FALSE
)
}
\arguments{
\item{multi_nucleotides_counts}{Multi-Nucleotide counts matrix.}

\item{samples}{name of the samples. This should match a rownames in multi_nucleotides_counts}

\item{freq}{boolean value; shall I display rates instead of counts?}

\item{xlabels}{boolean value; shall I display x labels?}
}
\value{
A ggplot2 object.
}
\description{
Plot Multi-Nucleotide Variants (MNVs) counts for a set of given patients.
}
\examples{
data(plot_data_examples)
counts = plot_data_examples[["patients.MNV.plot"]][["counts"]]
patients.MNV.plot(multi_nucleotides_counts=counts,samples=rownames(counts)[1:2])

}
