% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{patients.SBS.plot}
\alias{patients.SBS.plot}
\title{patients.SBS.plot}
\usage{
patients.SBS.plot(
  trinucleotides_counts,
  samples = rownames(trinucleotides_counts),
  freq = FALSE,
  xlabels = FALSE
)
}
\arguments{
\item{trinucleotides_counts}{trinucleotides counts matrix.}

\item{samples}{name of the samples. This should match a rownames in trinucleotides_counts.}

\item{freq}{boolean value; shall I display rates instead of counts?}

\item{xlabels}{boolean value; shall I display x labels?}
}
\value{
A ggplot2 object.
}
\description{
Plot Single Base Substitutions (SBS) counts for a set of given patients.
}
\examples{
data(plot_data_examples)
counts = plot_data_examples[["patients.SBS.plot"]][["counts"]]
patients.SBS.plot(trinucleotides_counts=counts,samples=rownames(counts)[1:2])

}
