% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{patientsIDPlot}
\alias{patientsIDPlot}
\title{patientsIDPlot}
\usage{
patientsIDPlot(
  id_data_counts,
  samples = rownames(id_data_counts),
  freq = FALSE,
  xlabels = FALSE
)
}
\arguments{
\item{id_data_counts}{Small Insertions and Deletions counts matrix.}

\item{samples}{Name of the samples. This should match a rownames in id_data_counts}

\item{freq}{Boolean value; shall I display rates instead of counts?}

\item{xlabels}{Boolean value; shall I display x labels?}
}
\value{
A ggplot2 object.
}
\description{
Plot Small Insertions and Deletions (ID) counts for a set of given patients.
}
\examples{
data(plot_data_examples)
counts <- plot_data_examples[['patients.ID.plot']][['counts']]
patientsIDPlot(id_data_counts=counts,samples=rownames(counts)[seq_len(2)])

}
