% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{plot_data_examples}
\alias{plot_data_examples}
\title{List data structure to run examples}
\format{
List data structure to run examples
}
\source{
List data structure to run examples.
}
\usage{
data(plot_data_examples)
}
\value{
List data structure to run examples
}
\description{
List data structure to run examples.
}
