% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RESOLVE-package.R
\docType{package}
\name{RESOLVE-package}
\alias{RESOLVE-package}
\alias{RESOLVE}
\title{An R package for the efficient analysis of mutational signatures from cancer genomes.}
\description{
This package provides a framework that allows for the efficient extraction and assignment of mutational signatures.
It implements a novel algorithm that enables: (i) the efficient extraction, (ii) exposure estimation, and
(iii) confidence assessment during the computational inference of mutational signatures.
}
\keyword{package}
