% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sbs_assignments}
\alias{sbs_assignments}
\title{SBS assignments for 560 breast tumors}
\format{
SBS assignments inferred by RESOLVE from 560 breast tumors published in Nik-Zainal, Serena, et al. (2016)
}
\source{
Nature. 2016 Jun 2;534(7605):47-54 (https://www.nature.com/articles/nature17676).
}
\usage{
data(sbs_assignments)
}
\value{
SBS assignments inferred by RESOLVE
}
\description{
SBS assignments from 560 breast tumors published in Nik-Zainal, Serena, et al. (2016).
}
