% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{signatures.CN.plot}
\alias{signatures.CN.plot}
\title{signatures.CN.plot}
\usage{
signatures.CN.plot(beta, useRowNames = FALSE, xlabels = FALSE)
}
\arguments{
\item{beta}{matrix with the inferred mutational signatures.}

\item{useRowNames}{boolean value; shall I use the rownames from beta as names for the signatures?}

\item{xlabels}{boolean value; shall I display x labels?}
}
\value{
A ggplot2 object.
}
\description{
Plot the inferred Copy Number (CN) mutational signatures.
}
