% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{signatures.CX.plot}
\alias{signatures.CX.plot}
\title{signatures.CX.plot}
\usage{
signatures.CX.plot(beta, useRowNames = FALSE, xlabels = FALSE)
}
\arguments{
\item{beta}{matrix with the inferred mutational signatures.}

\item{useRowNames}{boolean value; shall I use the rownames from beta as names for the signatures?}

\item{xlabels}{boolean value; shall I display x labels?}
}
\value{
A ggplot2 object.
}
\description{
Plot the inferred Copy Number (Reduced, CX) mutational signatures.
}
\examples{
data(plot_data_examples)
beta = plot_data_examples[["signatures.CX.plot"]][["beta"]]
signatures.CX.plot(beta=beta)

}
