% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{signaturesMNVPlot}
\alias{signaturesMNVPlot}
\title{signaturesMNVPlot}
\usage{
signaturesMNVPlot(beta, useRowNames = FALSE, xlabels = FALSE)
}
\arguments{
\item{beta}{Matrix with the inferred mutational signatures.}

\item{useRowNames}{Boolean value; shall I use the rownames from beta as names for the signatures?}

\item{xlabels}{Boolean value; shall I display x labels?}
}
\value{
A ggplot2 object.
}
\description{
Plot the inferred Multi-Nucleotide Variants (MNVs) mutational signatures.
}
\examples{
data(plot_data_examples)
beta <- plot_data_examples[['signatures.MNV.plot']][['beta']]
signaturesMNVPlot(beta=beta)

}
