% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ssm560_reduced}
\alias{ssm560_reduced}
\title{A reduced version of the point mutations for 560 breast tumors in the format compatible with the import function}
\format{
Reduced version of the counts of the point mutations in the format compatible with the import function
}
\source{
Nature. 2016 Jun 2;534(7605):47-54 (https://www.nature.com/articles/nature17676).
}
\usage{
data(ssm560_reduced)
}
\value{
Reduced version of the counts of point mutations for 560 tumors and 96 trinucleotides in the format compatible with the import function
}
\description{
Reduced version of the dataset of counts of the point mutations detected in 560 breast tumors published in Nik-Zainal, Serena, et al. (2016).
}
