% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.import.R
\name{getMNVCounts}
\alias{getMNVCounts}
\title{getMNVCounts}
\usage{
getMNVCounts(data)
}
\arguments{
\item{data}{a data.frame with variants having 6 columns: sample name, chromosome, start position, end position, ref, alt.}
}
\value{
A matrix with Multi-Nucleotide Variants (MNVs) counts per patient.
}
\description{
Create Multi-Nucleotide Variants (MNVs) counts matrix from input data.
}
\examples{
data(ssm560_reduced)
res <- getMNVCounts(data = ssm560_reduced)

}
