% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{groupsCXPlot}
\alias{groupsCXPlot}
\title{groupsCXPlot}
\usage{
groupsCXPlot(counts, groups, normalize = TRUE, xlabels = FALSE)
}
\arguments{
\item{counts}{matrix with Copy Number (Reduced, CX) counts data.}

\item{groups}{list where names are groups labels and elements are patients labels corresponding to rownames in counts.}

\item{normalize}{boolean value; shall I normalize observed counts?}

\item{xlabels}{boolean value; shall I display x labels?}
}
\value{
A ggplot2 object.
}
\description{
Plot observed Copy Number (Reduced, CX) counts for different groups of patients.
}
