% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resolve.R
\name{resolve.save}
\alias{resolve.save}
\title{Save a Trained RESOLVE Model}
\usage{
resolve.save(trainer, path)
}
\arguments{
\item{trainer}{A trained Trainer object (from resolve.train())}

\item{path}{Path to save checkpoint}
}
\description{
Save model checkpoint.
}
\examples{
\dontrun{
result <- resolve.train(dataset)
resolve.save(result, "model_checkpoint.pt")
}

}
