% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutational.signatures.discovery.R
\name{signaturesAssignment}
\alias{signaturesAssignment}
\title{signaturesAssignment}
\usage{
signaturesAssignment(
  x,
  beta,
  normalize_counts = FALSE,
  sparsify = TRUE,
  verbose = TRUE
)
}
\arguments{
\item{x}{counts matrix for a set of n patients and m categories. These can be, e.g., trinucleotides counts for n patients and 96 trinucleotides.}

\item{beta}{matrix of the discovered signatures to be used for the assignment.}

\item{normalize_counts}{if true, the input counts matrix x is normalized such that the patients have the same number of mutation.}

\item{sparsify}{boolean; Shall I perform regularization using LASSO?}

\item{verbose}{boolean; Shall I print information messages?}
}
\value{
A list with the discovered signatures. It includes 2 elements:
             alpha: matrix of the discovered exposure values.
             beta: matrix of the discovered signatures.
}
\description{
Perform the assignment of K somatic mutational signatures provided as input to samples given a set of observed counts x.
}
\examples{
data(background)
data(patients)
set.seed(12345)
beta = signaturesDecomposition(x = patients[1:3,], 
                               K = 3, 
                               background_signature = background, 
                               nmf_runs = 2, 
                               sparsify = FALSE, 
                               num_processes = 1)
set.seed(12345)
res = signaturesAssignment(x = patients[1:3,], beta = beta$beta[[1]], sparsify = FALSE)

}
