% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutational.signatures.discovery.R
\name{signaturesDecomposition}
\alias{signaturesDecomposition}
\title{signaturesDecomposition}
\usage{
signaturesDecomposition(
  x,
  K,
  background_signature = NULL,
  normalize_counts = FALSE,
  nmf_runs = 50,
  sparsify = TRUE,
  num_processes = Inf,
  seed = NULL,
  verbose = TRUE
)
}
\arguments{
\item{x}{counts matrix for a set of n patients and m categories. These can be, e.g., trinucleotides counts for n patients and 96 trinucleotides.}

\item{K}{either one value or a range of numeric values (each of them greater than 0) indicating the number of signatures to be considered.}

\item{background_signature}{background signature to be used.}

\item{normalize_counts}{if true, the input counts matrix x is normalized such that the patients have the same number of mutation.}

\item{nmf_runs}{number of iteration (minimum 1) of NMF to be performed for a robust estimation of beta.}

\item{sparsify}{boolean; Shall I perform regularization using LASSO?}

\item{num_processes}{Number of processes to be used during parallel execution. To execute in single process mode, 
this parameter needs to be set to either NA or NULL.}

\item{seed}{Seed for reproducibility.}

\item{verbose}{boolean; Shall I print information messages?}
}
\value{
A list with the discovered signatures and related rank measures. It includes 3 elements:
             alpha: list of matrices of the discovered exposure values for each possible rank in the range K.
             beta: list of matrices of the discovered signatures for each possible rank in the range K.
             measures: a data.frame containing the quality measures for each possible rank in the range K.
}
\description{
Perform signatures discovery and rank estimation for a range of K somatic mutational signatures given a set of observed counts x.
}
\examples{
data(background)
data(patients)
res = signaturesDecomposition(x = patients[1:5,], 
                              K = 3:4, 
                              background_signature = background, 
                              nmf_runs = 2, 
                              sparsify = FALSE, 
                              num_processes = NA, 
                              seed = 12345)

}
